Move header of game page to its own component
